Moved global ability system to dedicated folder
